+++ src/CriteriaFormatter.php
@@ -8,6 +8,7 @@ use Heptacom\HeptaConnect\Dataset\Base\TaggedCollection\TaggedStringCollection;
 use Heptacom\HeptaConnect\Dataset\Base\TaggedCollection\TagItem;
 use Heptacom\HeptaConnect\Package\Shopware6\Contract\Criteria;
 use Heptacom\HeptaConnect\Package\Shopware6\Contract\CriteriaFormatterInterface;
+use Heptacom\HeptaConnect\Package\Shopware6\Contract\FieldSortingCollection;
 
 final class CriteriaFormatter implements CriteriaFormatterInterface
 {
@@ -20,6 +21,7 @@ final class CriteriaFormatter implements CriteriaFormatterInterface
         $ids = $criteria->getIds();
         $term = $criteria->getTerm();
         $includes = $criteria->getIncludes();
+        $sort = $criteria->getSort();
 
         if ($limit !== null) {
             $result['limit'] = $limit;
@@ -45,6 +47,10 @@ final class CriteriaFormatter implements CriteriaFormatterInterface
             $result['includes'] = $this->getIncludeValues($includes);
         }
 
+        if ($sort !== null) {
+            $result['sort'] = $this->getSortValues($sort);
+        }
+
         return $result;
     }
 
@@ -58,6 +64,24 @@ final class CriteriaFormatter implements CriteriaFormatterInterface
         /** @var TagItem<string> $include */
         foreach ($includes as $include) {
             $result[$include->getTag()] = $include->getCollection()->asArray();
+        }
+
+        return $result;
+    }
+
+    /**
+     * @return list<array{field: string, order: string, naturalSorting: bool}>
+     */
+    private function getSortValues(FieldSortingCollection $sort): array
+    {
+        $result = [];
+
+        foreach ($sort as $instruction) {
+            $result[] = [
+                'field' => $instruction->getField(),
+                'order' => $instruction->getDirection(),
+                'naturalSorting' => $instruction->isNaturalSorting(),
+            ];
         }
 
         return $result;