+++ test/Integration/AdminApi/Utility/EntityClientTest.php
@@ -20,6 +20,8 @@ use Heptacom\HeptaConnect\Package\Shopware6\Test\Integration\AdminApi\Action\Abs
  * @covers \Heptacom\HeptaConnect\Package\Shopware6\Contract\Criteria
  * @covers \Heptacom\HeptaConnect\Package\Shopware6\Contract\Entity
  * @covers \Heptacom\HeptaConnect\Package\Shopware6\Contract\EntityCollection
+ * @covers \Heptacom\HeptaConnect\Package\Shopware6\Contract\FieldSorting
+ * @covers \Heptacom\HeptaConnect\Package\Shopware6\Contract\FieldSortingCollection
  * @covers \Heptacom\HeptaConnect\Package\Shopware6\CriteriaFormatter
  * @covers \Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Action\AbstractActionClient
  * @covers \Heptacom\HeptaConnect\Package\Shopware6\Http\AdminApi\Authentication\ApiConfiguration